refactor: split CLI and harden webhook boundaries
Reduce boundary indirection across CLI, flow wrappers, and webhook orchestration so behavior is easier to reason about and review. Add direct CLI and webhook handoff coverage to keep these refactors from hiding regressions.
