+++ .pre-commit-config.yaml
@@ -39,7 +39,7 @@ repos:
   # Python linting and formatting with Ruff
   - repo: https://github.com/astral-sh/ruff-pre-commit
     # Ruff version.
-    rev: v0.15.9
+    rev: v0.15.10
     hooks:
       # Run the linter.
       - id: ruff-check
@@ -49,7 +49,7 @@ repos:
 
   # Formatting for various file types with Prettier
   - repo: https://github.com/rbubley/mirrors-prettier  # Using rbubley mirror
-    rev: v3.8.1
+    rev: v3.8.2
     hooks:
       - id: prettier
         types_or: [html, css, scss, javascript, json]
@@ -82,7 +82,7 @@ repos:
 
   # Secret scanning - disabled due to Go version incompatibility in CI
   - repo: https://github.com/gitleaks/gitleaks
-    rev: v8.30.0
+    rev: v8.30.1
     hooks:
       - id: gitleaks
 
@@ -119,7 +119,7 @@ repos:
 
   # conventional commit messages. Allows for automated changelogs and versioning based on commits
   - repo: https://github.com/commitizen-tools/commitizen
-    rev: v4.13.9
+    rev: v4.13.10
     hooks:
       - id: commitizen
         stages: [commit-msg]
@@ -136,5 +136,5 @@ repos:
       - id: skylos
         name: skylos
         language: system
-        entry: uv run --with skylos skylos --danger --quality --secrets src/canvas_code_correction/
+        entry: uv run --with skylos skylos --json --danger --quality --secrets src/canvas_code_correction/
         pass_filenames: false
