+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "uv_build"
 [dependency-groups]
 dev = [
   "canvas-code-correction[ci,docs,test]",
-  "ipython>=9.11.0",
+  "ipython>=9.12.0",
   "poethepoet>=0.42.1",
   "commitizen>=4.13.9"
 ]
@@ -34,11 +34,11 @@ dependencies = [
   "typer>=0.24.1",
   "canvasapi>=3.5.0",
   "boto3>=1.42.70",
-  "fastapi>=0.135.1",
+  "fastapi>=0.135.2",
   "uvicorn[standard]>=0.42.0",
   "PyJWT>=2.12.0",
   "limits>=5.8.0",
-  "requests>=2.32.5",
+  "requests>=2.33.0",
   "rich>=14.3.3",
   "python-slugify>=8.0.4"
 ]
@@ -48,7 +48,7 @@ docs = [
   "httpx>=0.28.1",
   "mkdocs-mermaid2-plugin>=1.2.3",
   "mkdocs-glightbox>=0.5.2",
-  "zensical>=0.0.28"
+  "zensical>=0.0.29"
 ]
 test = [
   "pytest>=9.0.3",
@@ -57,9 +57,10 @@ test = [
   "pytest-cov>=7.1.0"
 ]
 ci = [
-  "prek>=0.3.6",
-  "pyrefly>=0.57.1",
-  "ruff>=0.15.7",
+  "prek>=0.3.8",
+  "pyrefly>=0.58.0",
+  "ruff>=0.15.8",
+  "skylos>=4.1.3",
   "smokeshow>=0.5.0"
 ]
 
@@ -76,13 +77,10 @@ version_provider = "pep621"
 
 [tool.poe.tasks]
 # Development workflow
-fmt = {cmd = "ruff format", help = "Formats Python code using Ruff."}
 lint = {cmd = "ruff check --fix", help = "Lints and fixes Python code using Ruff."}
-check = {cmd = "pyrefly check", help = "Performs static type checking with Pyrefly."}
 test = {cmd = "pytest", help = "Runs project tests using pytest."}
 all = {sequence = ["fmt", "lint", "check", "test"], help = "Runs all development checks: format, lint, type check, and tests."}
 clean = {cmd = "uvx pyclean . --debris", help = "Clean up Python cache files and test artifacts using pyclean."}
-"ci:fmt" = {cmd = "ruff format --check", help = "Checks if code is formatted without making changes (for CI)."}
 "ci:lint" = {cmd = "ruff check", help = "Lints code without applying fixes (for CI)."}
 # Project-specific tasks
 docker-publish = {sequence = ["_docker_build", "_docker_push"], help = "Build and push docker image for base runner"}
@@ -136,7 +134,9 @@ known-first-party = ["canvas_code_correction"]
 "tests/**/*.py" = ["S101", "D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107", "INP001", "PLR0913", "PLR2004", "ANN001", "ANN201", "ARG001", "FBT001", "PLR0915", "S108", "PLC0415", "N818", "SIM117", "RET504", "SLF001", "S105", "TC003", "TRY300", "T201"]
 # Ignore all linting issues in development scripts
 "scripts/**/*.py" = ["ALL"]
-"src/canvas_code_correction/cli.py" = ["PLR0913"]
+"src/canvas_code_correction/cli.py" = ["ANN", "ARG", "COM", "D", "E", "F", "FBT", "I", "N", "PLR", "UP"]
+"src/canvas_code_correction/cli_course.py" = ["ANN", "ARG", "COM", "D", "E", "F", "FBT", "I", "N", "PLR", "UP"]
+"src/canvas_code_correction/cli_system.py" = ["ANN", "ARG", "COM", "D", "E", "F", "FBT", "I", "N", "PLR", "UP"]
 
 [tool.uv.sources]
 canvas-code-correction = {workspace = true}
