+++ src/canvas_code_correction/flows/correction.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import tempfile
 from dataclasses import dataclass
 from pathlib import Path
-from typing import TYPE_CHECKING, Protocol, cast
+from typing import Protocol, cast
 
 from prefect import flow, task
 
@@ -13,25 +13,23 @@ from canvas_code_correction.clients.canvas_resources import (
     CanvasResources,
     build_canvas_resources,
 )
-from canvas_code_correction.collector import CanvasMetadataValue, ResultCollector
-from canvas_code_correction.runner import (
+from canvas_code_correction.config import Settings
+from canvas_code_correction.flows.collector import CanvasMetadataValue, ResultCollector
+from canvas_code_correction.flows.runner import (
     GraderConfig,
     GraderExecutor,
     create_default_grader_config,
 )
-from canvas_code_correction.uploader import (
+from canvas_code_correction.flows.uploader import (
     UploadConfig,
     UploadDetailValue,
     create_uploader_from_resources,
 )
-from canvas_code_correction.workspace import (
+from canvas_code_correction.flows.workspace import (
     WorkspaceConfig,
     WorkspacePaths,
     prepare_workspace,
 )
-
-if TYPE_CHECKING:
-    from canvas_code_correction.config import Settings
 
 
 @dataclass(frozen=True)
@@ -194,13 +192,6 @@ def fetch_submission_metadata(
 
     Prefect will handle retries and logging around this task.
     """
-    return _fetch_submission_metadata(resources, payload)
-
-
-def _fetch_submission_metadata(
-    resources: CanvasResources,
-    payload: CorrectSubmissionPayload,
-) -> SubmissionMetadata:
     assignment = resources.course.get_assignment(payload.assignment_id)
     submission = assignment.get_submission(
         payload.submission_id,
@@ -216,6 +207,13 @@ def _fetch_submission_metadata(
         assignment=_canvas_object_to_dict(assignment),
         submission=_canvas_object_to_dict(submission),
     )
+
+
+def _fetch_submission_metadata(
+    resources: CanvasResources,
+    payload: CorrectSubmissionPayload,
+) -> SubmissionMetadata:
+    return fetch_submission_metadata.fn(resources, payload)
 
 
 def _canvas_object_to_dict(obj: object) -> dict[str, CanvasMetadataValue]:
@@ -270,14 +268,6 @@ def download_submission_files(
     destination: Path,
 ) -> list[Path]:
     """Download submission attachments into the provided destination."""
-    return _download_submission_files(resources, payload, destination)
-
-
-def _download_submission_files(
-    resources: CanvasResources,
-    payload: CorrectSubmissionPayload,
-    destination: Path,
-) -> list[Path]:
     destination.mkdir(parents=True, exist_ok=True)
 
     assignment = resources.course.get_assignment(payload.assignment_id)
@@ -299,6 +289,14 @@ def _download_submission_files(
         downloaded_files.append(local_path)
 
     return downloaded_files
+
+
+def _download_submission_files(
+    resources: CanvasResources,
+    payload: CorrectSubmissionPayload,
+    destination: Path,
+) -> list[Path]:
+    return download_submission_files.fn(resources, payload, destination)
 
 
 @task
@@ -522,6 +520,7 @@ def correct_submission_flow(
     dry_run: bool = False,
 ) -> FlowArtifacts:
     """Prefect flow orchestrating the CCC correction stages."""
+    settings = Settings.model_validate(settings)
     resources = resources or build_canvas_resources(settings)
     metadata = fetch_submission_metadata(resources, payload)
